Make the linter in wasm mode happy

+++ flutter_ffi_plugin/example/native/hub/src/bridge/interface_web.rs
@@ -31,7 +31,7 @@ pub fn send_dart_signal_extern(
         None
     };
     let _ = catch_unwind(|| {
-        crate::messages::generated::handle_dart_signal(message_id as i32, message_bytes, blob);
+        crate::messages::generated::handle_dart_signal(message_id, message_bytes, blob);
     });
 }
